--- client/public/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-
-<script>
-    function getbooks(){
-        document.getElementById("output").innerHTML="";
-        fetch("http://openlibrary.org/search.json?q="+document.getElementById("input").value)
-        .then(a => a.json())
-        .then(response =>{
-            for(var i=0; i<10; i++){
-                document.getElementById("output").innerHTML+=
-                "<h2>"+response.docs[i].title+"</h2>"
-                    +response.docs[i].author_name[0]+
-                    "<br><img src='http://covers.openlibrary.org/b/isbn/"+response.docs[i].isbn[0]+"-L.jpg'><br>";
-            }
-        });
-    }
-</script>
-        <input id="input">        
-<button onclick="getbooks()">Get Books</button><br>
-<div id="output"></div>
