--- client/src/pages/About.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import NavBar from './NavBar';
-
-const HomePage = () => {
-  return (
-    <div>
-      <NavBar />
-      <h1>Home Page</h1>
-      <p>Welcome to the home page!</p>
-    </div>
-  );
-};
-
-export default HomePage;
