--- client/src/pages/NavBar.js
@@ -1,28 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-
-const NavBar = () => {
-  return (
-    <nav className='navbar'>
-      <ul>
-        <li>
-          <Link to="/">Home</Link>
-        </li>
-        <li>
-          <Link to="/about">About</Link>
-        </li>
-        <li>
-          <Link to="/contact">Events</Link>
-        </li>
-        <li>
-          <Link to="/shop">Shop</Link>
-        </li>
-        <li>
-          <Link to="/contact">Contact</Link>
-        </li>
-      </ul>
-    </nav>
-  );
-};
-
-export default NavBar;
