+++ myapp/src/pages/Events.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import NavBar from '../pages/NavBar';
+
+const HomePage = () => {
+  return (
+    <div>
+      <NavBar />
+      <h1 className='title-logo'>A Novel Idea</h1>
+      <p>We like big books and we cannot lie!</p>
+    </div>
+  );
+};
+
+export default HomePage;
